requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
distro==1.9.0
fal_client==0.13.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.13.0
msgpack==1.1.2
openai==2.21.0
pillow==12.1.1
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
PySimpleGUI==5.0.10
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
